+++ CHANGES.rst
@@ -6,6 +6,7 @@ Unreleased Changes
 
 * Document release process
 * Document how to run in ECS as a Scheduled Task
+* ``LocalCommand`` class - if ``script_source`` parameter is specified, instead of ignoring ``command`` send it as arguments to the downloaded script.
 
 0.2.0 (2017-11-30)
 ------------------